refactor: improve date parsing and error handling in BitcoinExchange; update RPN to use int64_t for calculations

+++ src/cpp09/ex01/RPN.hpp
@@ -12,7 +12,7 @@ class RPN
 {
 private:
     std::stack<int> container_;
-    int popHelper();
+    int64_t popHelper();
 
 public:
     RPN();
